feat: common out render cd material from trigger view and env trigger view
